refactor(web): adopt _components folder convention
Rename app/components to the Next.js private _components folder so
route-local pieces are clearly excluded from routing, and replace the
(global-not-found) route group with a _components/NotFound folder that
splits the 404 into composable Main/Title/Description pieces.

global-not-found.tsx now composes those pieces directly and the layout
and home page import from ./_components.
