--- apps/web/src/app/(global-not-found)/components/index.ts
@@ -1,1 +1,0 @@
-export * from './BackHomeButton';
