+++ apps/web/src/app/_components/NotFound/index.ts
@@ -1,0 +1,4 @@
+export * from './NotFoundMain';
+export * from './NotFoundTitle';
+export * from './NotFoundDescription';
+export * from './BackHomeButton';
